Modified and added new 'Point' class.

+++ GAME/src/entity/enemy/basicEnemy.java
@@ -2,10 +2,10 @@ package entity.enemy;
 
 public class basicEnemy extends Enemy {
     private static double   _hitPoint    =   40;
-    private static int      _speed       =   100;
+    private static double   _speed       =   100;
     private static int      _reward      =   5;
     private static double   _armor       =   2.5;
-    private static int      _damage      =   5;
+    private static double   _damage      =   5;
 
     public basicEnemy(int _X, int _Y) {
         super(_X, _Y);
